+++ zigar-runtime/src/accessors/all.js
@@ -1,7 +1,7 @@
 import { memberNames, MemberType } from '../constants.js';
 import { mixin } from '../environment.js';
 import { FALLBACK } from '../symbols.js';
-import { defineProperty, defineValue, usize } from '../utils.js';
+import { defineProperty, defineValue } from '../utils.js';
 
 // handle retrieval of accessors
 
@@ -29,69 +29,44 @@ export default mixin({
       names.push(`@${bitOffset}`);
     }
     const accessorName = access + names.join('');
+    let accessor = this.accessorCache.get(accessorName);
+    if (accessor) {
+      return accessor;
+    }
     // see if it's a built-in method of DataView
-    let accessor = DataView.prototype[accessorName];
+    accessor = DataView.prototype[accessorName];
+    if (!accessor) {
+      while (names.length > 0) {
+        const handlerName = `getAccessor${names.join('')}`;
+        if (accessor = this[handlerName]?.(access, member)) {
+          break;
+        }
+        names.pop();
+      }
+      /* c8 ignore start */
+      if (!accessor) {
+        throw new Error(`No accessor available: ${accessorName}`);
+      }
+      /* c8 ignore end */
+    }
     if (process.env.TARGET === 'node') {
       if (accessor && this.usingBufferFallback()) {
-        const thisEnv = this;
         const normal = accessor;
-        const getAddress = function(offset) {
-          const { buffer, byteOffset, byteLength } = this;
-          const base = buffer[FALLBACK];
-          if (base) {
-            if (offset < 0 || (offset + (bitSize >> 3)) > byteLength) {
-              throw new RangeError('Offset is outside the bounds of the DataView');
-            }
-            return base + usize(byteOffset + offset);
-          }
-        };
         accessor = (access === 'get')
         ? function(offset, littleEndian) {
-            const address = getAddress.call(this, offset);
-            if (address !== undefined) {
-              return thisEnv.getNumericValue(type, bitSize, address);
-            } else {
-              return normal.call(this, offset, littleEndian);
-            }
+            this[FALLBACK]?.(false, offset, byteSize);
+            return normal.call(this, offset, littleEndian);
           }
         : function(offset, value, littleEndian) {
-            const address = getAddress.call(this, offset);
-            if (address !== undefined) {
-              return thisEnv.setNumericValue(type, bitSize, address, value);
-            } else {
-              return normal.call(this, offset, value, littleEndian);
-            }
+            normal.call(this, offset, value, littleEndian);
+            this[FALLBACK]?.(true, offset, byteSize);
           };
       }
     }
-    if (accessor) {
-      return accessor;
+    if (!accessor.name) {
+      defineProperty(accessor, 'name', defineValue(accessorName));
     }
-    // check cache
-    accessor = this.accessorCache.get(accessorName);
-    if (accessor) {
-      return accessor;
-    }
-    while (names.length > 0) {
-      const handlerName = `getAccessor${names.join('')}`;
-      if (accessor = this[handlerName]?.(access, member)) {
-        break;
-      }
-      names.pop();
-    }
-    /* c8 ignore start */
-    if (!accessor) {
-      throw new Error(`No accessor available: ${accessorName}`);
-    }
-    /* c8 ignore end */
-    defineProperty(accessor, 'name', defineValue(accessorName));
     this.accessorCache.set(accessorName, accessor);
     return accessor;
   },
-  ...(process.env.TARGET === 'node' ? {
-    imports: {
-      getNumericValue: null,
-      setNumericValue: null,
-    },
-  } : undefined),
 });