+++ zigar-runtime/src/utils.js
@@ -1,5 +1,5 @@
 import { MemberType } from './constants.js';
-import { ENVIRONMENT, LENGTH, MEMORY, NO_CACHE, RESTORE, SIGNATURE } from './symbols.js';
+import { ENVIRONMENT, FALLBACK, LENGTH, MEMORY, NO_CACHE, RESTORE, SIGNATURE } from './symbols.js';
 
 export function defineProperty(object, name, descriptor) {
   if (descriptor) {
@@ -336,13 +336,22 @@ export function createView(size) {
 
 export function copyView(dest, src, offset = 0) {
   const destA = new Uint8Array(dest.buffer, dest.byteOffset, dest.byteLength);
+  if (process.env.TARGET === 'node') {
+    src[FALLBACK]?.(false, offset);
+  }
   const srcA = new Uint8Array(src.buffer, src.byteOffset, src.byteLength);
-  destA.set(srcA, offset);  
+  destA.set(srcA, offset);
+  if (process.env.TARGET === 'node') {
+    dest[FALLBACK]?.(true, offset);
+  }
 }
 
-export function clearView(dest, len = dest.byteLength, offset = 0) {
+export function clearView(dest, offset = 0, len = dest.byteLength - offset) {
   const destA = new Uint8Array(dest.buffer, dest.byteOffset, dest.byteLength);
-  destA.fill(0, offset, len);
+  destA.fill(0, offset, offset + len);
+  if (process.env.TARGET === 'node') {
+    dest[FALLBACK]?.(true, offset, len);
+  }
 }
 
 /* c8 ignore start */